Collapse Tier 1 harness returns into pipeline_tier1 module.
Introduce Tier1RunResult and a single builder for elapsed-time results
instead of six duplicated return paths in run.py.
