+++ packages/api/tests/test_inference_corpus_coverage.py
@@ -215,7 +215,7 @@ def test_expect_coverage_fails_without_solver_when_ground_truth_unavailable():
             "tests.inference_corpus.pipeline_preflight.extract_ground_truth_v1",
             return_value=unavailable_extraction,
         ),
-        patch("tests.inference_corpus.run.run_inference_with_artifacts") as run_inference,
+        patch("tests.inference_corpus.pipeline_tier1.run_inference_with_artifacts") as run_inference,
     ):
         result = run_manifest_case(coverage_required)
         run_inference.assert_not_called()
@@ -236,7 +236,7 @@ def test_available_ground_truth_action_not_in_catalog_out_of_search_without_expe
             "tests.inference_corpus.pipeline_preflight.extract_ground_truth_v1",
             return_value=forced_extraction,
         ),
-        patch("tests.inference_corpus.run.run_inference_with_artifacts") as run_inference,
+        patch("tests.inference_corpus.pipeline_tier1.run_inference_with_artifacts") as run_inference,
     ):
         result = run_manifest_case(no_expect_coverage)
         run_inference.assert_not_called()
